+++ cmd/update/cluster/command.go
@@ -26,6 +26,7 @@ Options:
   --name <cluster-name>             	Name of the cluster to update.
   --namespace <cluster-namespace>   	Namespace of the cluster.
   --release-version <release-version>   Update the cluster to a release version. The release version must be higher than the current release version.
+  --remove-schedule						Optionally: Remove the scheduled time for the cluster.
   --scheduled-time <scheduled-time>     Optionally: Scheduled time when cluster should be updated, time format 'YYYY-MM-DD HH:MM'.
   --provider <provider> 		Name of the provider.`
 
@@ -36,7 +37,10 @@ kubectl gs update cluster --help
 kubectl gs update cluster --name abcd1 --namespace my-org --release-version 16.1.0 --provider aws
 
 # Schedule cluster update
-kubectl gs update cluster --name abcd1 --namespace my-org --release-version 16.1.0 --scheduled-time "2022-01-01 10:00" --provider aws`
+kubectl gs update cluster --name abcd1 --namespace my-org --release-version 16.1.0 --scheduled-time "2022-01-01 10:00" --provider aws
+
+# Remove schedule for cluster update
+kubectl gs update cluster --name abcd1 --namespace my-org --provider aws --remove-schedule`
 )
 
 type Config struct {