+++ cmd/update/cluster/flag.go
@@ -9,6 +9,7 @@ import (
 const (
 	flagName           = "name"
 	flagReleaseVersion = "release-version"
+	flagRemoveSchedule = "remove-schedule"
 	flagScheduledTime  = "scheduled-time"
 	flagProvider       = "provider"
 )
@@ -18,6 +19,7 @@ type flag struct {
 	print          *genericclioptions.PrintFlags
 	Name           string
 	ReleaseVersion string
+	RemoveSchedule bool
 	ScheduledTime  string
 	Provider       string
 }
@@ -26,6 +28,8 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.Name, flagName, "", "Name of the cluster to update.")
 
 	cmd.Flags().StringVar(&f.ReleaseVersion, flagReleaseVersion, "", "Update the cluster to a release version. The release version must be higher than the current release version.")
+
+	cmd.Flags().BoolVar(&f.RemoveSchedule, flagRemoveSchedule, false, "Remove the schedule time for the cluster.")
 
 	cmd.Flags().StringVar(&f.ScheduledTime, flagScheduledTime, "", "Optionally: Scheduled time when cluster should be updated. The value has to be in RFC822 Format and UTC time zone.")